@@ -12,14 +12,14 @@
         }
     ],
     "require": {
-        "php": ">=7.0",
+        "php": ">=7.2",
         "guzzlehttp/guzzle": "^6.3",
         "illuminate/notifications": "~5.5 || ~6.0",
         "illuminate/support": "~5.5 || ~6.0"
     },
     "require-dev": {
-        "mockery/mockery": "^0.9.5",
-        "phpunit/phpunit": "5.*"
+        "mockery/mockery": "^1.2",
+        "phpunit/phpunit": "^8.0"
     },
     "autoload": {
         "psr-4": {
